+++ internal/infra/table/extractor_db2.go
@@ -15,6 +15,7 @@
 // You should have received a copy of the GNU General Public License
 // along with LINO.  If not, see <http://www.gnu.org/licenses/>.
 
+//go:build db2
 // +build db2
 
 package table
@@ -25,6 +26,7 @@ import (
 	// import db2 connector
 	_ "github.com/ibmdb/go_ibm_db"
 
+	"github.com/cgi-fr/lino/internal/infra/commonsql"
 	"github.com/cgi-fr/lino/pkg/table"
 )
 
@@ -38,10 +40,12 @@ type Db2ExtractorFactory struct{}
 
 // New return a Db2 extractor
 func (e *Db2ExtractorFactory) New(url string, schema string) table.Extractor {
-	return NewSQLExtractor(url, schema, Db2Dialect{})
+	return NewSQLExtractor(url, schema, Db2Dialect{commonsql.Db2Dialect{}})
 }
 
-type Db2Dialect struct{}
+type Db2Dialect struct {
+	commonsql.Dialect
+}
 
 func (d Db2Dialect) SQL(schema string) string {
 	SQL := `
@@ -67,4 +71,12 @@ func (d Db2Dialect) SQL(schema string) string {
 	}
 
 	return SQL
+}
+
+func (d MariadbDialect) Select(tableName string) string {
+	query := "SELECT * FROM "
+	query += tableName
+	query += " FETCH FIRST 0 ROWS ONLY"
+
+	return query
 }