+++ internal/infra/table/storage_yaml.go
@@ -99,11 +99,13 @@ func (s YAMLStorage) Store(tables []table.Table) *table.Error {
 		for _, rc := range r.Columns {
 			cols = append(cols, YAMLColumn{Name: rc.Name, Export: rc.Export, Import: rc.Import})
 		}
+
 		yml := YAMLTable{
 			Name:    r.Name,
 			Keys:    r.Keys,
 			Columns: cols,
 		}
+
 		list.Tables = append(list.Tables, yml)
 	}